Code Explaination
Upload project document, explain the the way of implement. Use multiple thread to find peak of numeral different rows because it may appear many peaks in the row and it is hard to find a efficient algorithm to find all peaks of a 2D matrix. Using multiple thread can shorten execution time such that it almost finish at the time the file is read completely.